--- app/src/main/java/com/example/umc_stepper/domain/model/response/comment_controller/CommentResponseItem.kt
@@ -1,10 +1,0 @@
-package com.example.umc_stepper.domain.model.response.comment_controller
-
-data class CommentResponseItem(
-    val commentId: Int=0,
-    val content: String="",
-    val dateTime: String="",
-    val memberName: String="",
-    val postId: Int=0,
-    val profileImage: String=""
-)
